check if plot needs resizing
adds a needsResizing flag to the magicplotsource which is set whenever a magicplotcomponent is resized. can be used to redraw any plots.

this situation was handled in the filter plot by storing a rectangle and comparing it to the one passed into drawPlot. seeing as the analyzer also needs redrawing (eg when audio is paused and the spectrum is frozen), i thought this might generalize the necessity.

for the oscilloscope this is useless, of course. but for any plots which can ever possibly be static, it seemed like it could be useful.